Add files via upload
Recorded violent crime in the UK has been on the rise since 2011, with significant differences between regions in recorded violent crime. Crime has always been a particular problem in Leeds compared to other major cities in the UK. According to a police report in December 2020, crime rates in Leeds are higher than the average for similar areas. In general, deprivation is associated with violent crime, and the more deprived an area is, the greater the likelihood of violent crime. However, common sense alone is not enough to prove the correlation between two variables. This project will use Leeds, a city with a very high crime rate, as the research site and conduct correlation analysis through the crime records of Leeds Police Station and the deprivation rankings of various districts in Leeds. to verify the correlation between deprivation and violent crime.

This repository contains a notebook used to study the correlation between deprivation and crime numbers.

The aim of the coding was to explore the correlation between deprivation and crime numbers in Leeds.

To rerun the contents of the notebook, 3 more data sets are needed:
1. Deprivation ranking data for England in 2019 can be obtained from "https://www.cdrc.ac.uk".
2. West Yorkshire Police crime data for April 2021 can be obtained from "https://www.police.uk"
3. Shapefile for Leeds, available from "https://geoportal.statistics.gov.uk"

The ipynb notebook used for this project is named "Assessment2".